moves mocks to be unexported for s3
