+++ instrumentation/github.com/aws/aws-sdk-go/service/s3/otels3/mocks.go
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// package mocks provides a mock interface to the S3 client
-package mocks
+package otels3
 
 import (
 	"fmt"
@@ -24,27 +23,27 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
 )
 
-type MockS3Client struct {
+type mockS3Client struct {
 	s3iface.S3API
 }
 
 // PutObjectWithContext implements the mocked S3 client successfully returning an empty PutObjectOutput type
-func (s *MockS3Client) PutObjectWithContext(ctx aws.Context, input *s3.PutObjectInput, opts ...request.Option) (*s3.PutObjectOutput, error) {
+func (s *mockS3Client) PutObjectWithContext(ctx aws.Context, input *s3.PutObjectInput, opts ...request.Option) (*s3.PutObjectOutput, error) {
 	return &s3.PutObjectOutput{}, nil
 }
 
 // GetObjectWithContext implements the mocked S3 client successfully returning an empty GetObjectOutput type
-func (s *MockS3Client) GetObjectWithContext(ctx aws.Context, input *s3.GetObjectInput, opts ...request.Option) (*s3.GetObjectOutput, error) {
+func (s *mockS3Client) GetObjectWithContext(ctx aws.Context, input *s3.GetObjectInput, opts ...request.Option) (*s3.GetObjectOutput, error) {
 	return &s3.GetObjectOutput{}, nil
 }
 
 // DeleteObjectWithContext implements the mocked S3 client successfully returning an empty DeleteObjectOutput type
-func (s *MockS3Client) DeleteObjectWithContext(ctx aws.Context, input *s3.DeleteObjectInput, opts ...request.Option) (*s3.DeleteObjectOutput, error) {
+func (s *mockS3Client) DeleteObjectWithContext(ctx aws.Context, input *s3.DeleteObjectInput, opts ...request.Option) (*s3.DeleteObjectOutput, error) {
 	return &s3.DeleteObjectOutput{}, nil
 }
 
 // DeleteObject implements the mocked S3 client successfully returning an empty DeleteObjectOutput type
-func (s *MockS3Client) DeleteObject(input *s3.DeleteObjectInput) (*s3.DeleteObjectOutput, error) {
+func (s *mockS3Client) DeleteObject(input *s3.DeleteObjectInput) (*s3.DeleteObjectOutput, error) {
 	fmt.Printf("UnInstrumentedMethod `DeleteObject` called")
 	return &s3.DeleteObjectOutput{}, nil
 }
